Add ability to import/export mappings with profile
The certprofile_show command now takes a --mappings-out flag that writes
mappings to a JSON-formatted file, and the same file format can be read
in with the --mappings-file parameter of certprofile_import and
certprofile_mod.

The exception handling here is a bit complicated, but the new
exc_callback functions should take care of making sure that state is
restored in the event of a failure. (And with the addition of an
exc_callback, overriding execute() is no longer necessary.)

https://fedorahosted.org/freeipa/ticket/4899